requirements.txt
streamlit
pandas
numpy
scikit-learn
scipy
matplotlib
seaborn
plotly
statsmodels
openpyxl
sweetviz
fpdf
PyPDF2
python-docx
missingno

# Machine Learning & Predictive Modeling
xgboost
lightgbm
prophet

# Clustering & Dimensionality Reduction
umap-learn

# Model Explainability
shap
